requirements.txt
aiohttp
PyYAML
nuitka
aiofiles
